+++ core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.30.0 [unreleased]
 
+- Add `ConnectionLimit::with_max_established` (see [PR 2137]).
+
 - Add `Keypair::to_protobuf_encoding` (see [PR 2142]).
 
 - Change `PublicKey::into_protobuf_encoding` to `PublicKey::to_protobuf_encoding` (see [PR 2145]).
@@ -12,6 +14,7 @@
 
 [PR 2145]: https://github.com/libp2p/rust-libp2p/pull/2145
 [PR 2142]: https://github.com/libp2p/rust-libp2p/pull/2142
+[PR 2137]: https://github.com/libp2p/rust-libp2p/pull/2137/
 
 # 0.29.0 [2021-07-12]